Code refactoring pcs 23 (#16)
* Fixed cost in form from Details

* Added the functionality in Details form of being able to switch between view and modification

* Details page and details config page is now only 1 page, modified attributes from the table Subscription of the database.

* Updated postdeployment script.

* Fixed obligatory field (cost) in deduction form, added format of currency in the field.

* Added back arrow for a better navegation experience.

+++ Planilla_UCR/Domain/Subscriptions/DTOs/SubscriptionDTO.cs
@@ -9,13 +9,13 @@ namespace Domain.Subscriptions.DTOs
         public String SubscriptionName { get; set; }
         public String ProviderName { get; set; }
         public String SubscriptionDescription { get; set; }
-        public int Cost { get; set; }
+        public double Cost { get; set; }
         public int TypeSubscription { get; set; }
         public int IsEnabled { get; set; }
 
         public SubscriptionDTO(String employerEmail, String projectName,
             String subscriptionName, String providerName, 
-            String subscriptionDescription, int cost, int typeSubscription, int isEnabled)
+            String subscriptionDescription, double cost, int typeSubscription, int isEnabled)
         {
             EmployerEmail = employerEmail;
             ProjectName = projectName;
